refactor(orchestrator): dedup provider-output + isChineseLocale helpers (Stage 2)
Four helpers had byte-identical copies across the forecast pipeline:
- stripCodeFences / readOutputSizeBytes / formatRemainingTimeoutMs were
  duplicated verbatim in provider-runtime.ts and full-pulse.ts -> new
  lib/provider-output.ts (single source).
- isChineseLocale had three copies (provider-runtime / full-pulse /
  market-pulse), differing only in the locale type alias (both resolve to
  SkillLocale) -> moved next to SkillLocale in config.ts.

Behavior-preserving: function bodies moved verbatim; the merged locale type
is identical (ResolvedProviderSkillSettings["locale"] === SkillLocale).
orchestrator typecheck + build green; full suite 832/832.
